H.Res. 218 Engrossed in House
Providing for consideration of the bill (H.R. 1473) making appropriations for the Department of Defense and the other departments and agencies of the Government for the fiscal year ending September 30, 2011, and for other purposes; providing for consideration of the concurrent resolution (H. Con. Res. 35) directing the Clerk of the House of Representatives to make a correction in the enrollment of H.R. 1473; and providing for consideration of the concurrent resolution (H. Con. Res. 36) directing the Clerk of the House of Representatives to make a correction in the enrollment of H.R. 1473.

Version:  Engrossed in House (eh)
Date:     2011-04-13
Congress: 112

Source: https://www.govinfo.gov/content/pkg/BILLS-112hres218eh/xml/BILLS-112hres218eh.xml

Sponsored-By: D000492
Roll-Call: House 112-1-259 2011-04-13 Passed 242-183
Roll-Call: House 112-1-260 2011-04-13 Passed 241-179