H.Res. 294 Reported in House
Providing for consideration of the resolution (H. Res. 292) declaring that the President shall not deploy, establish, or maintain the presence of units and members of the United States Armed Forces on the ground in Libya, and for other purposes, and providing for consideration of the concurrent resolution (H. Con. Res. 51) directing the President, pursuant to section 5(c) of the War Powers Resolution, to remove the United States Armed Forces from Libya.

Version:  Reported in House (rh)
Date:     2011-06-02
Congress: 112

Source: https://www.govinfo.gov/content/pkg/BILLS-112hres294rh/xml/BILLS-112hres294rh.xml

Sponsored-By: S001184