H.Res. 558 Engrossed in House
Directing the Clerk of the House of Representatives to provide a copy of the on-the-record portions of the audio backup file of the deposition of William R. Clemens that was conducted by the Committee on Oversight and Government Reform on February 5, 2008, to the prosecuting attorneys in the case of United States of America v. Clemens, No. 1:10-cr-00223-RBW (D.D.C.).

Version:  Engrossed in House (eh)
Date:     2012-02-17
Congress: 112

Source: https://www.govinfo.gov/content/pkg/BILLS-112hres558eh/xml/BILLS-112hres558eh.xml

Sponsored-By: D000492