H.Res. 788 Reported in House
Providing for consideration of the joint resolution (H.J. Res. 118) providing for congressional disapproval under chapter 8 of title 5, United States Code, of the rule submitted by the Office of Family Assistance of the Administration for Children and Families of the Department of Health and Human Services relating to waiver and expenditure authority under section 1115 of the Social Security Act (42 U.S.C. 1315) with respect to the Temporary Assistance for Needy Families program; providing for consideration of the bill (H.R. 3409) to limit the authority of the Secretary of the Interior to issue regulations before December 31, 2013, under the Surface Mining Control and Reclamation Act of 1977; and providing for proceedings during the period from September 22, 2012, through November 12, 2012.

Version:  Reported in House (rh)
Date:     2012-09-19
Congress: 112

Source: https://www.govinfo.gov/content/pkg/BILLS-112hres788rh/xml/BILLS-112hres788rh.xml

Sponsored-By: B001250